Extract requestobj to utils directory
